+++ R/wkt.R
@@ -42,6 +42,7 @@ polyline_wkt.sfencoded <- function(obj) {
   obj[[geomCol]] <- polyline_wkt(obj[[geomCol]])
 
   attr(obj[[geomCol]], "class") <- c("wkt_column", class(obj[[geomCol]] ) )
+  attr(obj, "class") <- c("wkt", setdiff(attr(obj, "class"), "sfencoded"))
   
   attr(obj, "encoded_column") <- NULL
   attr(obj, "wkt_column") <- geomCol
@@ -58,12 +59,64 @@ polyline_wkt.encoded_column <- function(obj) polyline_to_wkt(obj)
 #' @export
 polyline_wkt.default <- function(obj) stop(paste0("I was expecting an sfencoded object or an encoded_column"))
 
+#' @export
+`[.wkt` <- `[.sfencoded`
+
+
+#' @export
+print.wkt <- function(x, ... ){
+
+  if( is.null(attr(x, "wkt_column")) ) {
+    NextMethod()
+    return()
+  }
+
+  printwkt(x, ... )
+}
+
+printwkt <- function(x, ...) {
+
+  wkt <- attr(x, "wkt_column")
+
+  if(!is.null(wkt)) {
+    for (i in wkt) {
+      w <- as.character( x[[i]] )
+      w <- paste0(substr(w, 1, pmin(nchar(w), 30)), "...")
+      w <- stats::setNames(data.frame(w), i)
+      x[, i] <- w 
+    }
+  }
+
+  x <- removeWktClass(x)
+
+  print(x)
+  invisible(x)
+}
+
+removeWktClass <- function(x) { 
+  attr(x, "class") <- setdiff(class(x), "wkt")
+  return(x)
+}
+
+removeWktAttributes <- function(x) {
+  
+  wktCol <- attr(x, "wkt_column")
+  
+  if(!is.null(wktCol) && wktCol %in% names(x)) {
+    attr(x[[wktCol]], "class") <- NULL
+  }
+  
+  attr(x, "wkt_column") <- NULL
+  attr(x, "sfAttributes") <- NULL
+  return(x)
+}
 
 #' WKT Polyline
 #' 
 #' Converts well-known text into encoded polylines.
 #' 
-#' @param obj \code{sfencoded} object or \code{wkt_column} of well-known text
+#' @param obj \code{sfencoded} object, \code{wkt_column} or character vector
+#'  of well-known text
 #' 
 #' @return encoded polyline representation of geometries
 #' 
@@ -82,6 +135,17 @@ polyline_wkt.default <- function(obj) stop(paste0("I was expecting an sfencoded 
 #' ## convert well-known text back to polylines
 #' enc2 <- wkt_polyline(wkt)
 #' 
+#' ## data.frame column of well-known text
+#' df <- data.frame(
+#'   wkt = c("LINESTRING(0 0, 1 1, 1 2, 2 2)")
+#' )
+#' 
+#' df$polyline <- wkt_polyline(df$wkt)
+#' 
+#' ## use \code{unlist} to strip attributes
+#' df$polyline <- unlist(df$polyline)
+#' 
+#' 
 #' }
 #' 
 #' @details
@@ -92,7 +156,7 @@ polyline_wkt.default <- function(obj) stop(paste0("I was expecting an sfencoded 
 wkt_polyline <- function(obj) UseMethod("wkt_polyline")
 
 #' @export
-wkt_polyline.sfencoded <- function(obj) {
+wkt_polyline.wkt <- function(obj) {
   
   if(is.null(attr(obj, "wkt_column"))) stop("Can not find the wkt_column")
   
@@ -104,6 +168,8 @@ wkt_polyline.sfencoded <- function(obj) {
   
   attr(obj, "wkt_column") <- NULL
   
+  attr(obj, "class") <- c("sfencoded", setdiff(attr(obj, "class"), "wkt"))
+  
   attr(obj, "encoded_column") <- geomCol
   return(obj)
 }
@@ -112,4 +178,10 @@ wkt_polyline.sfencoded <- function(obj) {
 wkt_polyline.wkt_column <- function(obj) wkt_to_polyline(obj)
 
 #' @export
-wkt_polyline.default <- function(obj) stop(paste0("I was expecting an sfencoded object with a wkt_column"))
+wkt_polyline.character <- function(obj) wkt_to_polyline(obj)
+
+#' @export
+wkt_polyline.factor <- function(obj) wkt_to_polyline(obj)
+
+#' @export
+wkt_polyline.default <- function(obj) stop(paste0("I was expecting a charactor vector or a wkt object with a wkt_column"))